% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bmCredit}
\alias{bmCredit}
\title{Tiles Servers Attribution}
\usage{
bmCredit(x)
}
\arguments{
\item{x}{a character string of the url of the server.}
}
\value{
The attribution of the server.
}
\description{
Take a tiles server url and return its attribution.
}

