% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{bmServer}
\alias{bmServer}
\title{Tiles Servers URL}
\usage{
bmServer(x)
}
\arguments{
\item{x}{a character string of the name of the server.}
}
\value{
The url of the server.
}
\description{
Take a tiles server name (as returned by \code{\link{bmSource}}) and return its url.
}

