% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bmSource}
\alias{bmSource}
\title{Basemap Tiles Servers}
\usage{
bmSource(print.servers = TRUE)
}
\arguments{
\item{print.servers}{logical. Should the names of the servers be printed?}
}
\value{
Returns invisibly a matrix with servers names, urls and credits.
}
\description{
Print a list of tiles servers ready-to-use with \code{\link{basemap}}.
}
