% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{campsites}
\alias{campsites}
\title{French Campsites}
\format{a \code{SpatialPolygonsDataFrame} with geometries of the 96 french departements (epsg:4326) and 11 variables.
\itemize{
   \item DEP.CODE The code number of each department.
   \item DEP.NAME The name of each department.
   \item CHF.NAME The name of the main (administrative) city of each department.
   \item REGION.NAME The name of the administrative french region of each department.
   \item N.CAMPSITES The number of campsites.
   \item N.5, N.4, N.3, N.2, N.1 The number of campsites for each ranking categories (i.e. stars).
   \item PITCHES The number of camp pitches for each department.
 }}
\source{
\itemize{
   \item Institut National de l'Information Géographique et Forestière (2014)
   \item ATOUT FRANCE - Agence de développement touristique de la France (2014).
 }
}
\usage{
data(campsites)
}
\description{
This dataset gives the number of ranked campsites and the number of tent pitches for each department of metropolitan France.
}
\keyword{datasets}

