% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incLeaflet}
\alias{incData}
\alias{incEncoding}
\alias{incLeaflet}
\alias{initMap0}
\alias{initMap1}
\title{Generate chunks of HTML/JS code}
\usage{
incLeaflet(loc)

incEncoding(code)

incData(prefix)

initMap0(height, width)

initMap1(setView, setZoom)
}
\description{
These functions are used internally by \code{\link{writeMap}} to generate chunks of HTML and Javascript code.
}

