% Generated by roxygen2 (4.0.2): do not edit by hand
\name{col2hexa}
\alias{col2hexa}
\title{Convert colors to hexadecimal format}
\usage{
col2hexa(x, alpha.channel = FALSE, alpha = 1, charstring = TRUE)
}
\arguments{
\item{x}{a vector of any of the three kinds of \R color specifications.}

\item{alpha.channel}{logical. Sould an alpha channel included in the output?  Default is \code{FALSE}.}

\item{alpha}{a vector of numeric values in \eqn{[0, 1]}. Recycled if necessary.}

\item{charstring}{logical. Sould resulting elements be surrounded by quotation marks? Default is \code{TRUE}.}
}
\value{
A character vector of hexadecimal values.
}
\description{
This function converts any color of the \R system to hexadecimal format.
}
\seealso{
\code{\link[grDevices]{col2rgb}} for translating \R colors to RGB vectors.
}
\keyword{internal}

