% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{hotels}
\alias{hotels}
\title{French Hotels}
\format{a \code{SpatialPolygonsDataFrame} with geometries of the 96 french departements (epsg:4326) and 12 variables.
\itemize{
   \item DEP.CODE The code number of each department.
   \item DEP.NAME The name of each department.
   \item CHF.NAME The name of the main (administrative) city of each department.
   \item REGION.NAME The name of the french region (administrative) of each department.
   \item N.HOTELS The number of hotels.
   \item N.5, N.4, N.3, N.2, N.1 The number of hotels for each ranking categories (i.e. stars).
   \item ROOMS The number of hotel's rooms for each department.
   \item CAPACITY The total capacity (beds) for each department.
 }}
\source{
\itemize{
   \item Institut National de l'Information Géographique et Forestière (2014)
   \item ATOUT FRANCE - Agence de développement touristique de la France (2014).
 }
}
\usage{
data(hotels)
}
\description{
This dataset gives the number of hotels, number of rooms and capacity for each department of metropolitan France.
}
\keyword{datasets}

