% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coords2json}
\alias{class2var}
\alias{cleanDepsub}
\alias{coords2json}
\alias{holify}
\alias{pngasp}
\alias{polycoords}
\alias{polyholes}
\alias{polyorder}
\alias{proclines}
\alias{procpolys}
\alias{safeVar}
\alias{xvarnames}
\title{Internals functions}
\usage{
coords2json(x, type, holes = NULL, order = NULL)

holify(a, b)

polycoords(x)

polyholes(x)

polyorder(x)

proclines(x)

procpolys(x)

pngasp(x)

class2var(x)

xvarnames(x)

safeVar(x)

cleanDepsub(x)
}
\description{
These functions are used internally by \pkg{rleafmap}.
}

