% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genmeths.R
\name{print.splpoints}
\alias{print.splgrid}
\alias{print.splicons}
\alias{print.spllines}
\alias{print.splpoints}
\alias{print.splpolygons}
\title{Printing spl object}
\usage{
\method{print}{splpoints}(x, ...)

\method{print}{splicons}(x, ...)

\method{print}{spllines}(x, ...)

\method{print}{splpolygons}(x, ...)

\method{print}{splgrid}(x, ...)
}
\arguments{
\item{x}{an \code{spl} object.}

\item{...}{additional arguments. Not used.}
}
\description{
Print spl objects
}

