% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rleafmap-package.R
\docType{package}
\name{rleafmap}
\alias{rleafmap}
\title{rleafmap}
\description{
Display spatial data with interactive maps powered by the open-
source JavaScript library 'Leaflet' (see <https://leafletjs.com/>). Maps can be
rendered in a web browser or displayed in the HTML viewer pane of 'RStudio'.
This package is designed to be easy to use and can create complex maps with
vector and raster data, web served map tiles and interface elements.
}
