% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splayer.R
\name{spLayer.default}
\alias{spLayer.default}
\title{Define a Vector Data Layer}
\usage{
\method{spLayer}{default}(x, ...)
}
\arguments{
\item{x}{a spatial object as defined in the package \pkg{sp}.}

\item{...}{additional arguments to pass to the function.}
}
\description{
Define a Vector Data Layer
}

