% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spLayer.SpatialGridDataFrame}
\alias{spLayer.SpatialGridDataFrame}
\title{Define a Raster Data Layer}
\usage{
\method{spLayer}{SpatialGridDataFrame}(x, name = NULL, layer,
  cells.col = heat.colors(12), cells.alpha = 1, ...)
}
\arguments{
\item{name}{a character string to name the layer.}

\item{x}{a spatial object (see Details).}

\item{layer}{which layer to select?}

\item{cells.col}{a vector of any of the three kinds of \R color specifications giving a gradient to color the grid.}

\item{cells.alpha}{a vector of numeric values in \eqn{[0, 1]} setting grid opacity.}
}
\description{
\code{spLayer.SpatialGridDataFrame} defines a new data layer from an object \code{SpatialGridDataFrame}.
}

