\name{spLayer.SpatialLines}
\alias{spLayer.SpatialLines}
\title{Define a Vector Data Layer}
\usage{
\method{spLayer}{SpatialLines}(x, name = NULL, stroke = T, stroke.col = 1,
  stroke.lwd = 1, stroke.lty = 1, stroke.alpha = 1, fill = T,
  fill.col = 2, fill.alpha = 0.5, label = NULL, popup = "", ...)
}
\description{
Define a Vector Data Layer
}

