% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splayer.R
\name{spLayer.SpatialPolygons}
\alias{spLayer.SpatialPolygons}
\title{Define a Vector Data Layer}
\usage{
\method{spLayer}{SpatialPolygons}(
  x,
  name = NULL,
  stroke = TRUE,
  stroke.col = 1,
  stroke.lwd = 1,
  stroke.lty = -1,
  stroke.alpha = 1,
  fill = TRUE,
  fill.col = 2,
  fill.alpha = 0.5,
  label = NULL,
  popup = "",
  popup.rmd = FALSE,
  holes = FALSE,
  legend = NULL,
  ...
)
}
\arguments{
\item{x}{a spatial object (see Details).}

\item{name}{a character string to name the layer.}

\item{stroke}{logical. Should a stroke be drawn along lines and polygons?}

\item{stroke.col}{a vector of any of the three kinds of \R color specifications to set strokes color.}

\item{stroke.lwd}{a numerical vector to set strokes width.}

\item{stroke.lty}{a character vector that defines the strokes dash patterns (See Details).}

\item{stroke.alpha}{a vector of numeric values in \eqn{[0, 1]} setting strokes opacity.}

\item{fill}{logical. Should points and polygons be filled?}

\item{fill.col}{a vector of any of the three kinds of \R color specifications to set fill colors.}

\item{fill.alpha}{a vector of numeric values in \eqn{[0, 1]} setting fill opacity.}

\item{label}{a reserved argument (in development).}

\item{popup}{a character vector giving contents for popups. HTML tags are accepted.}

\item{popup.rmd}{a logical indicating whether the popups should be processed as R Markdown with \pkg{knitr}. Default \code{FALSE}.}

\item{holes}{a logical indicating whether to use the \code{hole} slots of the \code{SpatialPolygons} object.}

\item{legend}{a legend object created with \code{\link{layerLegend}}.}

\item{...}{additional arguments to pass to the function.}
}
\description{
Define a Vector Data Layer
}
