% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genmeths.R
\name{summary.basemap}
\alias{summary.basemap}
\alias{summary.splgrid}
\alias{summary.splicons}
\alias{summary.spllines}
\alias{summary.splpoints}
\alias{summary.splpolygons}
\alias{summary.ui}
\title{Summary of map elements}
\usage{
\method{summary}{basemap}(object, ...)

\method{summary}{splpoints}(object, ...)

\method{summary}{splicons}(object, ...)

\method{summary}{spllines}(object, ...)

\method{summary}{splpolygons}(object, ...)

\method{summary}{splgrid}(object, ...)

\method{summary}{ui}(object, ...)
}
\arguments{
\item{object}{a map layer, \code{basemap}, \code{spl} or \code{ui} object.}

\item{...}{additional arguments. Not used.}
}
\description{
Get a summary of a map element.
}

