% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toGeoJSON.R
\name{toGeoJSON}
\alias{toGeoJSON}
\title{Convert a spl object to GeoJSON format}
\usage{
toGeoJSON(x, lightjson = F)
}
\arguments{
\item{x}{a \code{spl} object.}

\item{lightjson}{logical. Should GeoJSON code size be reduced by supressing extra whitespace characters and rounding numeric values?}
}
\value{
A character string of GeoJSON formatted code.
}
\description{
This function is used internally by \code{\link{writeMap}} to convert a given \code{spl} object to GeoJSON format.
}
\seealso{
\code{\link[rgdal]{writeOGR}} and its driver "\code{GeoJSON}" for GeoJSON export.
}

