% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toJS}
\alias{toJS}
\title{Generate Leaflet JS code for a given layer}
\usage{
toJS(x, url = "")
}
\arguments{
\item{x}{a \code{spl} or \code{basemap} object.}

\item{url}{a character string giving the path for the raster files.}
}
\value{
A character string of JavaScript Code.
}
\description{
This function is used internally by \code{\link{writeMap}} to generate Leaflet JavaScript code for a given layer.
}

