% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Options settings for map interface}
\usage{
ui(
  zoom = c("topleft", "topright", "bottomleft", "bottomright", "none"),
  layers = c("none", "topright", "topleft", "bottomleft", "bottomright"),
  attrib = c("bottomright", "topleft", "topright", "bottomleft", "none"),
  attrib.text = ""
)
}
\arguments{
\item{zoom}{a character string indicating if and how should the zoom control be displayed.
This must be one of "\code{topleft}", "\code{topright}", "\code{bottomleft}", "\code{bottomright}", "\code{none}"}

\item{layers}{a character string indicating if and how should the layers control be displayed.
This must be one of "\code{topleft}", "\code{topright}", "\code{bottomleft}", "\code{bottomright}", "\code{none}"}

\item{attrib}{a character string indicating if and how should the attribution control be displayed.
This must be one of "\code{topleft}", "\code{topright}", "\code{bottomleft}", "\code{bottomright}", "\code{none}"}

\item{attrib.text}{a character string for additionnal credits. HTML tags are accepted.}
}
\value{
An object of class \code{ui} which can be directly given as \code{interface} argument of \code{\link{writeMap}}.
}
\description{
Allow the user to choose which interface elements are displayed on the map and their positions.
}
