% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{velov}
\alias{velov}
\title{Vélo'v stations}
\format{a \code{SpatialPointsDataFrame} with the location of the 349 Vélov stations (epsg:4326) and 5 variables.
\itemize{
   \item ID The ID of each station.
   \item NAME The name of each station.
   \item ADDRESS The address of each station.
   \item DISTRICT The name of the administrative french "commune".
   \item CAPACITY The capacity of each sation.
 }}
\source{
Grand Lyon (14/04/2014)
}
\usage{
data(velov)
}
\description{
Stations of the bicycle sharing system of the city of Lyon (France).
}
\keyword{datasets}

