% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bmCredit}
\alias{bmCredit}
\title{Tiles Servers Attribution}
\usage{
bmCredit(x)
}
\arguments{
\item{x}{a character string of the url of the server.}
}
\value{
The attribution of the server.
}
\description{
Take a tiles server url and return its attribution.
}

