% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spLayerControl}
\alias{spLayerControl}
\title{Testing user inputs}
\usage{
spLayerControl(name, size = 1, stroke = T, stroke.col = 1,
  stroke.lwd = 1, stroke.lty = 1, stroke.alpha = 1, fill = T,
  fill.col = 1, fill.alpha = 1, label = "", popup = "", holes = F)
}
\description{
This function tests arguments validity for the function \code{\link{spLayer}}.
}

