% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{uiJS}
\alias{uiJS}
\title{Generate user interface JS code}
\usage{
uiJS(interface, ar)
}
\arguments{
\item{interface}{an \code{ui} object created with \code{\link{ui}}.}

\item{ar}{a list of \code{basemap} and \code{spl} objects.}
}
\description{
This function is used internally by \code{\link{writeMap}} to generate JavaScript code
related to the user interface.
}
