% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{velov}
\alias{velov}
\title{Vélo'v stations}
\format{a \code{SpatialPointsDataFrame} with the location and name of the 349 Vélov stations (epsg:4326).}
\source{
OpenStreetMap (14/04/2014)
}
\usage{
data(velov)
}
\description{
Stations of the bicycle sharing system of the city of Lyon (France).
}
\keyword{datasets}

