% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rleafmap-package.R
\docType{data}
\name{velov}
\alias{velov}
\title{Velo'v stations}
\format{a \code{SpatialPointsDataFrame} with the location and name of the 349 Velov stations (epsg:4326).}
\source{
OpenStreetMap (14/04/2014)
}
\usage{
data(velov)
}
\description{
Stations of the bicycle sharing system of the city of Lyon (France).
}
\keyword{datasets}

